+++ app/scripts/api.js
@@ -93,6 +93,14 @@ const createApi = function api(context, pubSub) {
       },
 
       /**
+       * Set a key that must be down in order for the .on('wheel') event to be published.
+       * @param {number} keyCode - The keyCode. If null or undefined, the condition will be removed.
+       */
+      setWheelCallbackKeydownCondition(keyCode) {
+        self.setWheelCallbackKeydownCondition(keyCode);
+      },
+
+      /**
        * Get the currently set auth header
        */
       getAuthHeader() {
@@ -659,6 +667,7 @@ const createApi = function api(context, pubSub) {
        * hgv.off('rangeSelection', rangeListener);
        * hgv.off('viewConfig', viewConfigListener);
        * hgv.off('mouseMoveZoom', mmz);
+       * hgv.off('wheel', wheelListener);
        * hgv.off('createSVG');
        */
       off(event, listenerId, viewId) {
@@ -680,6 +689,11 @@ const createApi = function api(context, pubSub) {
 
           case 'mouseMoveZoom':
             apiPubSub.unsubscribe('mouseMoveZoom', callback);
+            break;
+
+          case 'wheel':
+            self.hasWheelCallback = false;
+            apiPubSub.unsubscribe('wheel', callback);
             break;
 
           case 'rangeSelection':
@@ -853,6 +867,9 @@ const createApi = function api(context, pubSub) {
        * const mmz = event => console.log('Moved', event);
        * hgv.on('mouseMoveZoom', mmz);
        *
+       * const wheelListener = event => console.log('Wheel', event);
+       * hgv.on('wheel', wheelListener);
+       *
        * hgv.on('createSVG', (svg) => {
        *    const circle = document.createElement('circle');
        *    circle.setAttribute('cx', 100);
@@ -877,6 +894,10 @@ const createApi = function api(context, pubSub) {
 
           case 'mouseMoveZoom':
             return apiPubSub.subscribe('mouseMoveZoom', callback);
+
+          case 'wheel':
+            self.hasWheelCallback = true;
+            return apiPubSub.subscribe('wheel', callback);
 
           case 'rangeSelection':
             return apiPubSub.subscribe('rangeSelection', callback);