+++ test/APITests.js
@@ -830,6 +830,59 @@ describe('API Tests', () => {
       });
     });
 
+    it('triggers on wheel events with a key down condition', done => {
+      [div, api] = createElementAndApi(
+        simple1dHorizontalVerticalAnd2dDataTrack
+      );
+      const hgc = api.getComponent();
+      waitForTilesLoaded(hgc, () => {
+        api.on('wheel', e => {
+          expect(e.origEvt.clientX).toEqual(30);
+          expect(e.origEvt.clientY).toEqual(40);
+          done();
+        });
+
+        const key = 'y';
+        const keyCode = 89;
+        api.setWheelCallbackKeydownCondition(keyCode);
+        const keydownEvent = new KeyboardEvent('keydown', {
+          key,
+          keyCode
+        });
+        const keyupEvent = new KeyboardEvent('keyup', {
+          key,
+          keyCode
+        });
+        const canvas = findCanvas(div);
+        // The wheel event that we expect to be emitted before keydown,
+        // and therefore not passed to the callback.
+        const wheelEventBad = {
+          clientX: 10,
+          clientY: 20,
+          forwarded: true,
+          target: canvas,
+          nativeEvent: undefined,
+          stopPropagation: () => {},
+          preventDefault: () => {}
+        };
+        // The wheel event that we expect to be emitted after keydown.
+        const wheelEventGood = {
+          clientX: 30,
+          clientY: 40,
+          forwarded: true,
+          target: canvas,
+          nativeEvent: undefined,
+          stopPropagation: () => {},
+          preventDefault: () => {}
+        };
+        // Simulate the wheel and keyboard events.
+        hgc.wheelHandler(wheelEventBad);
+        hgc.keyDownHandler(keydownEvent);
+        hgc.wheelHandler(wheelEventGood);
+        hgc.keyUpHandler(keyupEvent);
+      });
+    });
+
     afterEach(() => {
       api.destroy();
       removeDiv(div);